learn page got updated also single side page and search added to both pages
